Apply Jules remediations for switch_port and sensor switch_port
